requirements.txt
flask
python-dotenv
openai
werkzeug
requests
